Latest fimware V1.2.2.34
mines comes with V1.2.2.20
this new one fix some issues and is working properly follow the instructions:
Upgrade by using the touch screen control:
When the machine is turned on normally, insert the U disk with the 【Upgrade Folder】into the USB port of the printer panel (Diagram 1-1), and wait for the U disk to be read successfully (Diagram 1-2, on the page of selecting files to print, check whether the file and folder information of the U disk can be read. If so, the U disk is mounted successfully.), then click【Settings】→【Information】on the main interface, click the upgrade button arrow on the information page (Diagram 1-3), and click confirm follow the prompt steps, and then upgrade the software. After the upgrade is successful, the machine will automatically restart. After machine restarting, please check whether the firmware version number has changed.